Renamed subkinect to kinect
