+++ dashboard/src/app.js
@@ -3,6 +3,6 @@ import ROSLIB from 'roslib';
 import $ from 'cash-dom';
 import route from './route';
 import gamestate from './subscribers/gamestate';
-import kinect from './subscribers/subkinect';
+import kinect from './subscribers/kinect';
 
 route(window.location.pathname, true);
