+++ dashboard/src/subscribers/kinect.js
