+++ dashboard/views/index.html
@@ -20,7 +20,7 @@
       <div class="uk-width-expand command-view command-current" name="gamestate" hidden>
         {% include 'subscriber.html' %}
       </div>
-      <div class="uk-width-expand command-view command-current" name="kinect" hidden>
+      <div class="uk-width-expand command-view" name="kinect" hidden>
         {% include 'kinect_subscriber.html' %}
       </div>
     </div>
